+++ src/components/CustomSpinner.js
@@ -1,11 +1,11 @@
 import { Spinner } from 'reactstrap';
 
-const CustomSpinner = () => {
+const CustomSpinner = ({label='Estamos buscando lo que nos pediste...'}) => {
 
     return (
         <div className='mt-5 mb-2'>
             <Spinner color="warning" className='mb-2'></Spinner>
-            <h1 className='lead'>Estamos procesando lo que nos pediste...</h1>
+            <h1 className='lead'>{label}</h1>
         </div>
     );
 }